get rid of CMSDEPRECATED_X warnings in RecoLocalTracker/SiPhase2VectorHitBuilder
